kobot/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
fonttools==4.39.3
frozendict==2.3.7
html5lib==1.1
idna==3.4
importlib-resources==5.12.0
install==1.3.5
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
multitasking==0.0.11
mysqlclient==2.1.1
numpy==1.24.3
packaging==23.1
pandas==1.4.4
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyupbit==0.2.33
QuantStats==0.0.59
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tabulate==0.9.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
websockets==11.0.2
yfinance==0.2.18
zipp==3.15.0
